changes requested in pull request: get_dummies problem + moving files around
